add monte carlo acquisition
