+++ bloptools/bayesian/digestion.py
@@ -1,4 +1,3 @@
 def default_digestion_function(db, uid):
     products = db[uid].table(fill=True)
-    print(products)
     return products
